% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Function responsible for initializing and running the REXoplanets shiny application.}
\usage{
app(..., run = TRUE)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::runApp()}.}

\item{run}{If true, runs the application. If false, returns an app object.}
}
\value{
App object.
}
\description{
Function responsible for initializing and running the REXoplanets shiny application.
}
