% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_esi.R
\name{calculate_esi}
\alias{calculate_esi}
\title{Calculate the Earth Similarity Index (ESI)}
\usage{
calculate_esi(pl_rade, pl_insol)
}
\arguments{
\item{pl_rade}{Numeric. Planetary radius in Earth radii.}

\item{pl_insol}{Numeric. Stellar flux in Earth units.}
}
\value{
Numeric. Earth Similarity Index (ESI).
}
\description{
Calculate the Earth Similarity Index (ESI)
}
\details{
The function calculates the ESI based on stellar flux and planetary radius.

ESI (Earth Similarity Index) is a characterization of how similar a planetary-mass object
or natural satellite is to earth. It was designed to be a scale from zero to one,
with Earth having a value of 1.
}
\examples{
calculate_esi(1, 1)        # Earth:   1
calculate_esi(0.53, 0.43)  # Mars:    0.753
calculate_esi(11.2, 0.037) # Jupiter: 0.24

}
\references{
Schulze-Makuch, D., Méndez, A., Fairén, A. G., von Paris, P., Turse, C., Boyer, G.,
Davila, A. F., Resendes de Sousa António, M., Irwin, L. N., and Catling, D. (2011)
A Two-Tiered Approach to Assess the Habitability of Exoplanets. Astrobiology 11(10): 1041-1052.
}
