% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-star-goldilocks-zone.R
\name{calculate_star_goldilocks_zone}
\alias{calculate_star_goldilocks_zone}
\title{Calculate star's Goldilocks zone}
\usage{
calculate_star_goldilocks_zone(st_lum, log_lum = TRUE)
}
\arguments{
\item{st_lum}{A numeric Stellar luminosity value (log10(L/Lsun) or linear).}

\item{log_lum}{A logical value. If \code{TRUE} assumes \code{st_lum} is logarithmic. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of 2 elements
where first is Goldilocks zone inner radius and second is the outer radius.
}
\description{
Calculates star's Goldilocks zone inner and outer radius,
based on star's luminosity.
}
\examples{
calculate_star_goldilocks_zone(0) # goldilocks zone for sun, with logarithmic units
calculate_star_goldilocks_zone(1, log_lum = FALSE) # goldilocks zone for sun, with linear units

}
