% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{closest_50_exoplanets}
\alias{closest_50_exoplanets}
\title{closest_50_exoplanets}
\format{
A data frame with 683 columns and 50 rows.
}
\source{
\url{https://exoplanetarchive.ipac.caltech.edu/}
}
\usage{
closest_50_exoplanets
}
\description{
A sample dataset of 50 exoplanets closest to earth,
taken from table \verb{Planetary Systems Composite Parameters},
including their names, discovery methods,
and other relevant information.
}
\keyword{datasets}
