% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_labels}
\alias{col_labels}
\title{col_labels
A collection of Exoplanets archive columns names and their respective labels / comments.
Currently only supports \code{ps} and \code{pscomppars} tables}
\format{
A nested list.
}
\source{
\url{https://exoplanetarchive.ipac.caltech.edu/docs/API_PS_columns.html#columns}
}
\usage{
col_labels
}
\description{
col_labels
A collection of Exoplanets archive columns names and their respective labels / comments.
Currently only supports \code{ps} and \code{pscomppars} tables
}
\examples{
col_labels[["ps"]][["pl_name"]]

}
\keyword{datasets}
