% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reactable_part.R
\name{.create_reactable_part}
\alias{.create_reactable_part}
\title{Internal helper for breaking up larger
\code{reactable} tables into a single row table}
\usage{
.create_reactable_part(output, id, data, col_selector)
}
\description{
Internal helper for breaking up larger
\code{reactable} tables into a single row table
}
\keyword{internal}
