% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_value_box_grid.R
\name{.create_value_box_grid}
\alias{.create_value_box_grid}
\title{Internal helper for creating grid of \code{valueboxes}
for displaying information.}
\usage{
.create_value_box_grid(df)
}
\description{
Internal helper for creating grid of \code{valueboxes}
for displaying information.
}
\keyword{internal}
