% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exoplanets_col_labels}
\alias{exoplanets_col_labels}
\title{exoplanets_col_labels
A collection of Exoplanets archive columns names and their respective labels / comments.
Currently only supports \code{ps} and \code{pscomppars} tables}
\format{
named vectors nested in a list.
}
\source{
\url{https://exoplanetarchive.ipac.caltech.edu/docs/API_PS_columns.html#columns}
}
\usage{
exoplanets_col_labels
}
\description{
exoplanets_col_labels
A collection of Exoplanets archive columns names and their respective labels / comments.
Currently only supports \code{ps} and \code{pscomppars} tables
}
\examples{
col_labels[["ps"]][["pl_name"]]

}
\keyword{datasets}
