% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_star_spectral_type.R
\name{get_star_spectral_type}
\alias{get_star_spectral_type}
\title{Get spectral type of a star}
\usage{
get_star_spectral_type(st_teff)
}
\arguments{
\item{st_teff}{Numeric. effectrive stellar temperature (in K)}
}
\value{
Character. spectral type classification with a letter:
\itemize{
\item \code{M}: Mercury-like planets (< 0.22 Earth masses)
\item \code{K}: Earth-like planets (0.22–2.2 Earth masses)
\item \code{G}: Super-Earths (2.2–22 Earth masses)
\item \code{F}: Neptune-like planets (22–127 Earth masses)
\item \code{A}: Jupiter-like giants (127–4450 Earth masses)
\item \code{B}: Degenerate-matter/brown dwarf-like objects (>= 4450 Earth masses)
\item \code{O}: Degenerate-matter/brown dwarf-like objects (>= 4450 Earth masses)
}
}
\description{
The function takes in effectrive stellar temperature (in K),
and return letter for the star's spectral type based on Morgan–Keenan (MK) system.
}
\examples{
get_star_spectral_type(5778)

}
