% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_planet_details.R
\name{module_planet_details}
\alias{module_planet_details}
\alias{planet_details_ui}
\alias{planet_details_server}
\title{Module Displaying Planet Details}
\usage{
planet_details_ui(id)

planet_details_server(id, planet_info)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{planet_info}{A reactive expression returning a data frame with information about the planet.}
}
\value{
A Shiny UI object.
}
\description{
Module Displaying Planet Details
}
\details{
This module provides a dynamically refreshing table component
for displaying basic information about a planet.
}
\section{Functions}{
\itemize{
\item \code{planet_details_ui()}: UI function for the planet details module.

\item \code{planet_details_server()}: server function for the planet details module.

}}
