% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_star_systems.R
\name{module_star_systems}
\alias{module_star_systems}
\alias{star_systems_ui}
\alias{star_systems_server}
\title{Star Systems Module}
\usage{
star_systems_ui(id)

star_systems_server(id, data)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{data}{A reactive expression that returns a data frame containing star system data.}
}
\value{
A Shiny UI object.

A Shiny server module.
}
\description{
Star Systems Module
}
\details{
This module provides functionality for exploring star systems and their planets.
Provides a handy UI for selecting a star system. The star is visualized alongside its planets and
their orbits. The display provides basic information about the system itself, as well as
particular planets. A legend is available for the plot.
}
\section{Functions}{
\itemize{
\item \code{star_systems_ui()}: UI function for the module.

\item \code{star_systems_server()}: Server function for the module.

}}
