% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_system_info.R
\name{module_system_info}
\alias{module_system_info}
\alias{system_info_ui}
\alias{system_info_server}
\title{Module Displaying Star System Info}
\usage{
system_info_ui(id)

system_info_server(id, system_info)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{system_info}{A reactive expression returning a data frame with information about the system.}
}
\value{
A Shiny UI object.
}
\description{
Module Displaying Star System Info
}
\details{
This module provides a dynamically refreshing table component
for displaying basic information about a star system.
}
\section{Functions}{
\itemize{
\item \code{system_info_ui()}: UI function for the system info module.

\item \code{system_info_server()}: server function for the system info module.

}}
