% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_system_plot_settings.R
\name{module_system_plot_settings}
\alias{module_system_plot_settings}
\alias{system_plot_settings_ui}
\alias{system_plot_settings_server}
\title{System Plot Settings Module}
\usage{
system_plot_settings_ui(id)

system_plot_settings_server(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A Shiny UI object.

A reactive list object containing bool value
whether to show the habitable zone and plot legend.
}
\description{
System Plot Settings Module
}
\details{
This module provides a series of controls to customize
the visible elements of a star system map
}
\section{Functions}{
\itemize{
\item \code{system_plot_settings_ui()}: UI function for the plot settings module.

\item \code{system_plot_settings_server()}: function for the search module.

}}
