% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_test.R
\name{module_test}
\alias{module_test}
\alias{test_ui}
\alias{test_server}
\title{Test module}
\usage{
test_ui(id)

test_server(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A Shiny UI object.

A Shiny server module.
}
\description{
Test module
}
\details{
This is just a test module for reference for further developement.
}
\section{Functions}{
\itemize{
\item \code{test_ui()}: UI function for the module.

\item \code{test_server()}: Server function for the module.

}}
