% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_visualize_star_system.R
\name{module_visualize_star_system}
\alias{module_visualize_star_system}
\title{Module Visualizing Star Systems}
\description{
Module Visualizing Star Systems
}
\details{
This module provides a dynamically refreshing plot component
for visualizing star systems based on data provided
}
