% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ps_colnames}
\alias{ps_colnames}
\title{ps_colnames
A collection of Planetary Systems table database column names
and their respective labels / comments.}
\format{
A named list of 299 elements.
}
\source{
\url{https://exoplanetarchive.ipac.caltech.edu/docs/API_PS_columns.html#columns}
}
\usage{
ps_colnames
}
\description{
ps_colnames
A collection of Planetary Systems table database column names
and their respective labels / comments.
}
\keyword{datasets}
