% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pscomppars_colnames}
\alias{pscomppars_colnames}
\title{pscomppars_colnames
A collection of Planetary Systems Composite Parameters table database column names
and their respective labels / comments.}
\format{
A named list of 326 elements.
}
\source{
\url{https://exoplanetarchive.ipac.caltech.edu/docs/API_PS_columns.html#columns}
}
\usage{
pscomppars_colnames
}
\description{
pscomppars_colnames
A collection of Planetary Systems Composite Parameters table database column names
and their respective labels / comments.
}
\keyword{datasets}
