% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rexoplanets_sample_data}
\alias{rexoplanets_sample_data}
\title{rexoplanets_sample_data}
\format{
A data frame with 50 columns and 81 rows.
}
\source{
\url{https://exoplanetarchive.ipac.caltech.edu/}
}
\usage{
rexoplanets_sample_data
}
\description{
A sample dataset of exoplanets, including their names, discovery methods,
and other relevant information.
}
\keyword{datasets}
