% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_star_occurrences}
\alias{summarize_star_occurrences}
\title{Summarize star occurrences}
\usage{
summarize_star_occurrences(data)
}
\arguments{
\item{data}{A data frame with KOI data. Must contain \code{kepoi_name} column.}
}
\value{
A data frame containing:
- \code{Star} column with star ID.
- \code{Count} column with number of planets.
}
\description{
Summarize star occurrences
}
\details{
The function takes in KOI data frame and summarizes the number of planets that appear in a
given dataset for each star.
}
