% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{TAP_URL}
\alias{TAP_URL}
\title{TAP_URL}
\format{
A string.
}
\source{
https://exoplanetarchive.ipac.caltech.edu/docs/TAP/usingTAP.html#bp
}
\usage{
TAP_URL
}
\description{
Link for NASA's exoplanets TAP service.
}
\keyword{internal}
