% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_test.R
\name{test_ui}
\alias{test_ui}
\alias{test_server}
\title{This is a test moduel for boilerplate purposes.}
\usage{
test_ui(id)

test_server(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A Shiny UI element.

A Shiny server function.
}
\description{
This is a test moduel for boilerplate purposes.

This is a test server function for the test module.
}
