% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_ps_table.R
\name{trim_ps_table}
\alias{trim_ps_table}
\title{Trim planetary systems table}
\usage{
trim_ps_table(data)
}
\arguments{
\item{data}{A data frame with planetary systems data.}
}
\value{
A data frame containing:
\itemize{
\item objectid - Object ID
\item pl_name – Planet name
\item hostname – Host star name
\item sy_dist – Distance to the system (parsecs)
\item pl_rade – Planetary radius (Earth radii)
\item pl_bmasse – Planetary mass (Earth masses)
\item pl_orbper – Orbital period (days)
\item pl_orbsmax – Semi-major axis (AU)
\item pl_insol – Incident stellar flux (Earth units)
\item st_teff – Stellar effective temperature (K)
\item st_rad – Stellar radius (Solar radii)
\item st_mass – Stellar mass (Solar masses)
\item st_lum – Stellar luminosity (log10 L/Lsun)
\item pl_eqt – Planetary equilibrium temperature (K)
\item pl_dens – Planetary density
\item discoverymethod – Discovery method
\item disc_year – Year of discovery
\item sy_snum – Number of stars in system
\item sy_pnum – Number of planets in system
\item pl_tsystemref – System parameter reference
\item pl_refname – Main reference publication
}
}
\description{
Trim planetary systems table
}
\details{
The function takes in a planetary systems data frame,
and trims it to include only 20 most important columns.
}
