% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Function responsible for initializing and running the REXoplanets shiny application.}
\usage{
app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\description{
Function responsible for initializing and running the REXoplanets shiny application.
}
