% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_esi.R
\name{calculate_esi}
\alias{calculate_esi}
\title{Calculate the Earth SImilarity Index (ESI)}
\usage{
calculate_esi(data)
}
\arguments{
\item{data}{A data frame with data from \code{ps} or \code{pscomppars} table.
It must contain columns: \code{st_lum}, \code{pl_orbsmax} and \code{pl_rade}.}
}
\value{
A data frame containing:
- \code{objectid} column with object ID.
- \code{esi} column with the ESI
}
\description{
Calculate the Earth SImilarity Index (ESI)
}
\details{
The function takes in data frame and calculates the esi for each entry.
}
