% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_esi.R
\name{calculate_esi}
\alias{calculate_esi}
\title{Calculate the Earth Similarity Index (ESI)}
\usage{
calculate_esi(..., weights = c(0.57, 0.7))
}
\arguments{
\item{...}{any number of numeric parameters representing the planet's characteristics.}

\item{weights}{Numeric vector of weights for each parameter. Default is c(0.57, 0.7),
assuming The first parameter is the radius, and the second is the stellar flux.}
}
\value{
Numeric. Earth Similarity Index (ESI).
}
\description{
Calculate the Earth Similarity Index (ESI)
}
\details{
The function calculates the ESI based for any parameter or number of parameters.
By default, it uses the weights of 0.57 for radius and 0.7 for stellar flux.

ESI (Earth Similarity Index) is a characterization of how similar a planetary-mass object
or natural satellite is to earth. It was designed to be a scale from zero to one,
with Earth having a value of 1.
}
\examples{
# ESI for radius and flux, returns 1 (Earth is perfectly similar to itself)
calculate_esi(1, 1)
# Mars approximation using radius ≈ 0.532 and flux ≈ 0.43, returns: ~0.812
calculate_esi(0.532, 0.43)
# Custom 3-parameter ESI (e.g. radius, flux, temperature)
calculate_esi(1.1, 1.2, 288, weights = c(0.5, 0.3, 0.2))

}
\references{
Schulze-Makuch, D., Méndez, A., Fairén, A. G., von Paris, P., Turse, C., Boyer, G.,
Davila, A. F., Resendes de Sousa António, M., Irwin, L. N., and Catling, D. (2011)
A Two-Tiered Approach to Assess the Habitability of Exoplanets. Astrobiology 11(10): 1041-1052.
}
