% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stellar_flux.R
\name{calculate_stellar_flux}
\alias{calculate_stellar_flux}
\title{Calculate stellar flux value}
\usage{
calculate_stellar_flux(data, log = TRUE, unit = "relative")
}
\arguments{
\item{data}{A data frame containing \code{objectid}, \code{st_lum}, and \code{pl_orbsmax}.}

\item{unit}{Character. Either \code{"relative"} (default) or \code{"wm2"} to convert to W/m².}

\item{log_lum}{Logical. If TRUE, assumes \code{st_lum} is in log10(L / L☉). Defaults to TRUE.}
}
\value{
A data frame with columns \code{objectid} and \code{stellar_flux}.
}
\description{
Calculate stellar flux value
}
\details{
This function calculates the stellar flux for each entry in the input data frame.
It assumes luminosity is either logarithmic (log10 of L/Lsun) or linear, and optionally
converts flux to absolute units (W/m²) if requested.

Stellar flux is the amount of energy from a star that reaches a given area per unit time.
}
